requirements.txt
sphinx_click
myst-parser
loguru
jmespath
netaddr
pytz
boto3
botocore
certifi
chardet
Click==7.0
colorclass
idna
jmespath
prompt-toolkit==2.0.10
Pygments
python-dateutil
PyYAML
regex
requests>=2.23.0
s3transfer
six
termcolor
terminaltables
tqdm
urllib3
wcwidth
loguru